+++ MatchingEngine/src/java/OrderRequest/ReplaceOrderRequest.java
@@ -5,8 +5,8 @@ import common.UserRequestType;
 public class ReplaceOrderRequest extends UserRequest {
     private final Long orderID;
     private final Integer quantity;
-    private final Long price;
-    public ReplaceOrderRequest(long orderID, Integer quantity, Long price) {
+    private final Double price;
+    public ReplaceOrderRequest(long orderID, Integer quantity, Double price) {
         super(UserRequestType.REPLACE_ORDER);
         this.orderID = orderID;
         this.quantity = quantity;
@@ -19,7 +19,7 @@ public class ReplaceOrderRequest extends UserRequest {
     public Integer getQuantity() {
         return quantity;
     }
-    public Long getPrice() {
+    public Double getPrice() {
         return price;
     }
 }
